feat: configurations beans
The first class is a JacksonMapper bean used to create a custom configuration for JSON input/output
on API endpoints.
The second is a ResourceBundle bean for get the message properties used for
validations, exceptions and custom messages with responses.
The thrid party enables the auditor
component for persist auditable fields in the DB such as date of creation and modification. Finally
the fourth is a SecurityConfig implementation allowing use authentication, and authorization of
users